requirements.txt
Django==4.2.10
django-reversion==5.0.5
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangochannelsrestframework==1.1.0
coreapi==2.3.3
drf_yasg==1.21.7
django-rest-swagger==2.2.0
django-reverse-admin==2.9.6
celery==5.3.6
